requirements.txt
gpiozero
lgpio
pigpio
numpy
opencv-python-headless
ultralytics
picamera2
pyyaml
